% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_github_branch.R
\name{create_github_branch}
\alias{create_github_branch}
\title{Create GitHub branch}
\usage{
create_github_branch(new_branch, sha, gh_user, gh_repo)
}
\arguments{
\item{new_branch}{name of the branch to create}

\item{sha}{SHA from which to create the branch. If left empty, the latest
commit from the default branch will be used}

\item{gh_user}{Name of GitHub user/org where the repo is hosted}

\item{gh_repo}{Name of GitHub repo}
}
\description{
Create GitHub branch
}
