% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_head.R
\name{get_default_head}
\alias{get_default_head}
\title{Get default HEAD}
\usage{
get_default_head(gh_user, gh_repo)
}
\arguments{
\item{gh_user}{Name of GitHub user/org where the repo is hosted}

\item{gh_repo}{Name of GitHub repo}
}
\description{
Get last commit from a GitHub default branch
}
\examples{
get_default_head("epiforecasts", "covid19-forecast-hub-europe")

}
